+++ R/transfer_centroids.R
@@ -1,0 +1,177 @@
+
+
+#   __________________ #< 45425d2ee7b52e512217af9ce04ee05b ># __________________
+#   Transfer centroids                                                      ####
+
+
+#' @title Transfer centroids from one data frame to another
+#' @description
+#'  \Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
+#'
+#'  Given two \code{data.frames} with the same columns (and groupings),
+#'  transfer the centroids from one to the other.
+#'
+#'  This is commonly used to restore the centroids after transforming the columns.
+#' @author Ludvig Renbo Olsen, \email{r-pkgs@@ludvigolsen.dk}
+#' @param to_data \code{data.frame}.
+#'
+#'  Existing \code{`dplyr`} groups are ignored. Specify in \code{`group_cols`} instead.
+#' @param from_data \code{data.frame} with the same columns (and groupings) as \code{`to_data`}.
+#'
+#'  Existing \code{`dplyr`} groups are ignored. Specify in \code{`group_cols`} instead.
+#' @param cols Names of numeric columns to transfer centroids to.
+#'  Must exist in both \code{`to_data`} and \code{`from_data`}.
+#' @param group_cols Names of grouping columns.
+#' @export
+#' @family clustering functions
+#' @return The \code{`to_data`} \code{data.frame} with the
+#'  centroids from the \code{`from_data`} \code{data.frame}.
+#' @examples
+#' \donttest{
+#' # Attach packages
+#' library(rearrr)
+#' library(dplyr)
+#'
+#' # Set seed
+#' set.seed(1)
+#'
+#' # Create a data frame
+#' df <- data.frame(
+#'   "x" = runif(20),
+#'   "y" = runif(20),
+#'   "g" = c(1, 1, 1, 1, 1,
+#'           2, 2, 2, 2, 2,
+#'           3, 3, 3, 3, 3,
+#'           4, 4, 4, 4, 4)
+#' )
+#'
+#' # Create another data frame with different x and y values
+#' df2 <- df
+#' df2$x <- runif(20)
+#' df2$y <- runif(20)
+#'
+#' # Check centroids before transfer
+#'
+#' df %>%
+#'   dplyr::group_by(g) %>%
+#'   dplyr::summarize_all(mean)
+#'
+#' df2 %>%
+#'   dplyr::group_by(g) %>%
+#'   dplyr::summarize_all(mean)
+#'
+#' # Now let's transfer the centroids from df to df2
+#'
+#' df3 <- transfer_centroids(
+#'   to_data = df2,
+#'   from_data = df,
+#'   cols = c("x", "y"),
+#'   group_cols = "g"
+#' )
+#'
+#' # Check that the transfer gave us the same centroids as df
+#' df3 %>%
+#'   dplyr::group_by(g) %>%
+#'   dplyr::summarize_all(mean)
+#'
+#' }
+transfer_centroids <- function(to_data,
+                               from_data,
+                               cols,
+                               group_cols = NULL) {
+  # Check arguments ####
+  assert_collection <- checkmate::makeAssertCollection()
+  checkmate::assert_data_frame(to_data,
+                               min.cols = 1,
+                               min.rows = 1,
+                               add = assert_collection)
+  checkmate::assert_data_frame(from_data,
+                               min.cols = 1,
+                               min.rows = 1,
+                               add = assert_collection)
+  checkmate::assert_character(cols,
+                              any.missing = FALSE,
+                              min.len = 1,
+                              add = assert_collection)
+  checkmate::assert_character(
+    group_cols,
+    any.missing = FALSE,
+    min.len = 1,
+    null.ok = TRUE,
+    add = assert_collection
+  )
+  checkmate::reportAssertions(assert_collection)
+  if (length(intersect(colnames(to_data), colnames(from_data))) != length(colnames(from_data))) {
+    assert_collection$push("'to_data' and 'from_data' must have the exact same columns.")
+  }
+  checkmate::reportAssertions(assert_collection)
+  # End of argument checks ####
+
+  # Ungroup data frames
+  to_data <- dplyr::ungroup(to_data)
+  from_data <- dplyr::ungroup(from_data)
+
+  # Group data frames if specified
+  if (!is.null(group_cols)) {
+    to_data <- dplyr::group_by(to_data,!!!rlang::syms(group_cols))
+    from_data <-
+      dplyr::group_by(from_data,!!!rlang::syms(group_cols))
+  }
+
+  # Find from centroids
+  from_centroids <- from_data %>%
+    dplyr::select(c(group_cols, cols)) %>%
+    dplyr::summarise_all(mean)
+
+  # Find centroids in the new data
+  to_centroids <- to_data %>%
+    dplyr::select(c(group_cols, cols)) %>%
+    dplyr::summarise_all(mean)
+
+  # Extract summarized group columns
+  from_group_columns <- from_centroids %>%
+    dplyr::ungroup() %>%
+    dplyr::select(group_cols)
+
+  # Extract summarized group columns
+  to_group_columns <- to_centroids %>%
+    dplyr::ungroup() %>%
+    dplyr::select(group_cols)
+
+  # Make sure the group columns are the same in
+  # both summaries
+  if (!dplyr::all_equal(from_group_columns,
+                        to_group_columns,
+                        ignore_row_order = FALSE)) {
+    stop("The summarized group columns from the two datasets are not equal.")
+  }
+
+  # Remove group columns
+  from_centroids <- from_centroids %>%
+    dplyr::ungroup() %>%
+    dplyr::select(cols)
+
+  # Remove group columns
+  to_centroids <- to_centroids %>%
+    dplyr::ungroup() %>%
+    dplyr::select(cols)
+
+  # How much to move centroids per dimension
+  distances <- from_centroids - to_centroids
+
+  # Group by group columns
+  distances <- dplyr::bind_cols(from_group_columns, distances) %>%
+    dplyr::group_by(!!!rlang::syms(group_cols))
+
+  # Move centroids!
+  to_data[, cols] <- purrr::map2_dfr(
+    .x = split(x = to_data, f = dplyr::group_indices(to_data)),
+    .y = split(x = distances, f = dplyr::group_indices(distances)),
+    .f = ~ {
+      .x[, cols, drop = FALSE] + .y[rep(1, nrow(.x)), cols, drop = FALSE]
+    }
+  )
+
+  to_data
+
+}