+++ NEWS.md
@@ -1,11 +1,9 @@
 
 # rearrr 0.0.0.9000
 
-* Adds `expand_values()` for expanding values around an origin in any number of dimensions. (Mutator)
+* Please welcome `rearrr`! :-)
 
-* Adds `rotate2d()` for rotating values around an origin. (Mutator)
-
-* Adds `flip_values()` for flipping values around a center value. (Mutator)
+## New rearrangers
 
 * Adds `center_max()` for centering the highest value with values decreasing around it. (Rearranger)
 
@@ -21,4 +19,22 @@
 
 * Adds `closest_to()` and `furthest_from()` for arranging values by how far they are from a target value / index. (Rearranger)
 
-* Please welcome `rearrr`! :-)
+## New mutators
+
+* Adds `cluster_groups()` for moving data points into clusters around their group centroid. (Mutator)
+
+* Adds `expand_values()` for expanding values around an origin in any number of dimensions. (Mutator)
+
+* Adds `rotate2d()` for rotating values around an origin. (Mutator)
+
+* Adds `flip_values()` for flipping values around a center value. (Mutator)
+
+## New helpers
+
+* Adds `centroid()` for finding the centroid from multiple vectors (dimensions). (Helper)
+
+* Adds `transfer_centroids()` for transfering centroids from one `data.frame` to another. (Helper)
+
+* Adds `min_max_scale()` for scaling values to a specified range. (Helper)
+
+* Adds `create_origin_fn()` for creating a function for finding origin coordinates (like `centroid()`). (Helper)